Removed old explore summaries code
